refactor(tauri): extract the native-theme decisions into their own module
Unit 11c of the ADR-035 epic, after 11a (pending update) and 11b (context
menus). `lib.rs` 7813 -> 6678; `native_theme.rs` holds the 21 items of the
cluster plus `theme_pref_tests`, which moves with them as a descendant so
it keeps reaching private items through `use super::*`.

Verbatim: 1135 lines, 9 differing, and all 9 are the `pub(crate)` the move
requires -- checked line-by-line against `git show HEAD:lib.rs`, not by
reading. Six of the nine are what `#[tauri::command]` forces (a wrapper
cannot be more visible than its argument and return types). The other
three -- `AppMode`, `HighContrast`, `AppModeOutcome` -- are reached by
`win_app_mode.rs` through `crate::` paths, which now name this module.

That last group is the one worth watching. `win_app_mode.rs` is
`#![cfg(target_os = "windows")]` and cfg-stripping runs before name
resolution, so getting it wrong compiles clean on macOS and Linux and
fails only on the Windows CI leg. The same blind spot covers the cluster's
own two bare `win_app_mode::` calls, which resolved only while this code
sat at the crate root. The `generate_handler!` entries become
module-qualified, matching `pending_update::` and `context_menu::`; the
wire contract is untouched because `#[tauri::command]` names are not
module-qualified.

The guard is the interesting half. `native-theme-claims.test.ts` reads
Rust as text and 8 of its 10 specs went red on the move -- the file
working, since each asserts it found something before asserting anything
about it. Re-pointing it at `native_theme.rs` would re-arm the same break
for 11d-11f, and concatenating every source file would let its several
first-hit `.match()` calls silently return another module's occurrence,
which is worse than finding nothing because it still reports a pass. So
the module is located by a construct that IS the feature, through a new
`rustSourceDefining` that asserts exactly one file matches. 11a's
disk-derived walk moved to `tests/docs/rust-sources.ts` and both callers
share it.

The guard then caught this commit's own widening: `pub(crate)` sits
between the attribute list and the item, and three `rename_all` checks
were anchored on those being adjacent.

7 of 7 mutations bite, including a second file defining `set_native_theme`.
Renaming `native_theme.rs` stays green on purpose -- it breaks no claim
here, and a guard that fails on a harmless rename is noise its next reader
learns to route around. That is stated in the control spec rather than
left to be rediscovered.
