+++ docs/plans/2026-08-24-ai-assisted-maintainability-remediation.md
@@ -89,7 +89,7 @@ Do not implement this document as one change.
 | 10c | Extract rail content/review coordination | Not started | — |
 | 11a | Extract Tauri pending-update marker | Merged | #1630 (merged 2026-08-26); pure move, verified line-by-line with zero lines not found verbatim (`lib.rs` -710/+5). `check_for_update_now` deliberately stayed: its body calls a private crate-root function four unrelated sites use, so moving it would have widened visibility OUTSIDE this cluster. Visibility is `pub(crate)`, not `pub` -- `mod pending_update;` is bare, so nothing is reachable from `src-tauri/tests/*.rs` either way. **The guard was widened to a disk-derived scan of `src-tauri/src`, not re-pointed at a fixed pair** -- a two-file list would have reproduced the same bug one extraction later, so 11b--11f are covered by construction. Widening a scan is where a guard becomes zero-of-zero (#1399), so each derivation step got its own control: a positive control on the walk (>10 files, both named), and the `ROUTED_ELSEWHERE` exemption must now be earned in a file the scan actually reached. **Two holes found by battery and review, not by reading.** A TEST was satisfying the production-routing claim -- `pending_update_tests` calls `surface_pending_update_hint_with(CODE_...)`, so rewriting the real call site to a literal left the file green; that hole PREDATES the extraction. Closed with `#[cfg(test)]` stripping -- which review then showed was itself defeated by text it already scans: `b"{ not json"` is an unbalanced brace in a byte-string literal, so the counter ran off the end, correct today ONLY because that test module is last in the file. The next item appended after it would have vanished from the scan silently. Literals are now skipped and an unbalanced block THROWS rather than truncating. **11 of 11 mutations bite.** `cargo fmt` deliberately not run (repo is not rustfmt-clean; it reformatted 16 untouched files). First full-gate run flaked on `apply-malformed.test.ts`'s parallel-rename race (#1599 area, load-dependent, identical commit green on rerun) -- recorded, not dismissed. `/code-review high` not run -- **merge policy changed mid-programme to /simplify + review-pr**, both of which ran |
 | 11b | Extract Tauri context-menu specifications | Merged | PR #1638 (merged 2026-08-26); verbatim move of 822 non-blank lines, 9 differing and all deliberate: seven `pub(crate)` prefixes, plus the `#[allow(dead_code)]` on `has_selection` (false -- four live read sites) and the prefix check becoming a named predicate. **The plan calls 11b the one cluster with no external guard test, and that is confirmed**: a repo-wide search found nothing reading `lib.rs` as text for a context-menu construct. So the move takes something with it that colocation was providing for free -- the `ctx:` prefix separating these ids from the tray's `MENU_*` ids was convention, visible while both id spaces lived in one file, asserted by nothing. `context_menu_id_space_tests` pins both directions over an id set derived by RUNNING every builder across its whole (all-bool) input space, with a positive control on the sweep. **Two holes closed that predate the unit**: deleting a name from `generate_handler!` or repointing `.on_menu_event` compiles clean with all 285 Rust tests green while `invoke()` starts rejecting in the WebView, so `tauri-command-registration-claims.test.ts` derives BOTH sides from source and fails loud on any `invoke()` argument it cannot resolve -- three real call sites pass a module constant, and a blind spot that reports itself is a gap while one that does not is a false pass. **14 of 15 mutations bite.** The survivor is named in the code: `forward_context_menu_event`'s emit is window-scoped so a second window cannot receive another's action, and flattening it to `app.emit(...)` is green everywhere -- reaching it needs a live `AppHandle` and a second window, so it is documented rather than covered by a source-text assertion that would read stronger than it is. **A method fix, because it nearly cost that finding**: an interrupted battery run left a mutation in the tree, the next run's backup adopted it as the baseline, and that row reported BITES when it does not; the harness now refuses to start, and fails after finishing, unless both suites are green. /simplify and review-pr converged on one real defect -- inserting the new predicate between `forward_context_menu_event` and its doc comment silently reassigned the doc to the predicate, since a blank line is what separates `///` blocks. `/code-review high` not run -- **merge policy changed mid-programme to /simplify + review-pr**, both of which ran |
-| 11c | Extract Tauri native-theme decisions | Not started | — |
+| 11c | Extract Tauri native-theme decisions | In review | Verbatim move of 1135 lines into `native_theme.rs` (`lib.rs` 7813 → 6678), **9 differing lines and all 9 are the `pub(crate)` the move requires** — checked line-by-line against `git show HEAD:lib.rs`, not by reading. **Plan review paid for itself: three of the four defects it found are invisible on this machine.** `win_app_mode.rs` reaches `AppMode`/`HighContrast`/`AppModeOutcome` by `crate::` path (23 sites) and is `#![cfg(target_os = "windows")]`, so my reference count read clean only because it never looked outside `lib.rs`, and the break would have surfaced on the Windows CI leg alone; the cluster's own two `win_app_mode::` calls are bare paths that resolve only at the crate root, same blind spot. The third — `generate_handler!` entries must become module-qualified — I had asserted the opposite and cited 11b as evidence, when 11b is evidence *against* it. The fourth was an off-by-seven on the test module's end line that would have swallowed the next module's doc comment. **The guard was the interesting part.** `native-theme-claims.test.ts` reads Rust as text and 8 of its 10 specs went red on the move — the file working, since each asserts it found something first. Re-pointing it at `native_theme.rs` would re-arm the same break for 11d–11f, and a concatenated corpus would let its several first-hit `.match()` calls silently return another module's occurrence. So the module is **located by a construct that IS the feature**, via a new `rustSourceDefining` that asserts exactly one file matches; 11a's disk-derived walk moved to `tests/docs/rust-sources.ts` and both files now share it. **The guard then caught my own widening**: `pub(crate)` sits between the attribute list and the item, and three `rename_all` checks anchored on adjacency. **7 of 7 mutations bite**, including a second file defining `set_native_theme`. Renaming `native_theme.rs` stays GREEN **on purpose** — it breaks no claim here, and a guard that fails on a harmless rename is noise; that is stated in the control spec rather than left to be rediscovered. `cargo fmt` deliberately not run |
 | 11d | Extract Tauri Cowork commands | Not started | — |
 | 11e | Extract Tauri sidecar lifecycle | Not started | — |
 | 11f | Reduce Tauri startup helpers; retain `lib.rs::run` root | Not started | — |
